Move arch-specific code out of virt.c
